graph, store: Move fulltext types to graph::schema
